Refactor: move all the action related actions to the `action` module.
