+++ lib/init.js
@@ -75,38 +75,6 @@ const readConfigFiles = ( configFiles ) => {
 	return unionConfig;
 };
 
-const readActionFile = ( actionFile ) => {
-	try {
-		const actionFunc = require( actionFile );
-
-		return actionFunc;
-	} catch ( error ) {
-		return false;
-	}
-};
-
-const readActionFiles = ( actionFiles ) => {
-	const actions = [];
-
-	// FIXME: can we make this smarter without the relative path?
-	const actionDir = '../actions/';
-	const localActionDir = '../local-actions/';
-
-	for ( const actionFile of actionFiles ) {
-		const result = readActionFile( localActionDir + actionFile ) ||
-						readActionFile( actionDir + actionFile );
-
-		if ( ! result ) {
-			console.error( 'A fatal error has occured when reading the action: ', actionFile );
-			return process.exit( -1 );
-		}
-
-		actions.push( result );
-	}
-
-	return actions;
-};
-
 const initialize = async ( config ) => {
 	const {
 		browser: browserSlug,
@@ -156,7 +124,6 @@ const configLocalStorage = async ( page, entries ) => {
 module.exports = {
 	configLocalStorage,
 	readConfigFiles,
-	readActionFiles,
 	initialize,
 	mergeConfig,
 };
